requirements.txt
future
emoji==1.7.0
lxml
wget
gtts
beautifulsoup4
requests
python-telegram-bot==13.14
SQLAlchemy==1.4.44
psycopg2-binary
feedparser
faker
webcolors
hachoir
pykeyboard
pynewtonmath
aiofiles
ffmpeg-python
spongemock
zalgo-text
geopy
nltk
psutil
aiohttp
Pillow
CurrencyConverter
googletrans==2.4.0
google-trans-new
jikanpy
speedtest-cli
coffeehouse==2.2.6
regex
bleach
wikipedia
markdown2
telethon
telegraph
heroku3
spamwatch
alphabet_detector
pyrate-limiter
cachetools
ujson
pretty_errors
pygithub
TgCrypto
pyrogram==1.4.16 
PyDictionary
git+https://github.com/Sorrybabe/downloader
youtube_search_python==1.6.6
youtube_search
lyrics_extractor
asyncio
cloudscraper
dateparser
pymongo
dnspython
secureme
apscheduler
emoji-country-flag
countryinfo
html2text
bs4
redis
pySmartDL
bing_image_downloader
search_engine_parser
pytz
jikanpy
tswift
python-arq
envparse
python-dotenv
better_profanity
textblob 
nudepy
nekos.py
motor
gpytranslate
fontTools
opencv-python-headless
odmantic
tzlocal
pendulum
fuzzysearch
img2pdf
multicolorcaptcha
git+https://github.com/python-telegram-bot/ptbcontrib.git
bs4
humanize
redis==4.3.5
aiohttp>=2.2.5
pytimeparse
pyYAML>=5.1.2
pyaml
aiohttp
nekos.py
html_telegraph_poster
